test/mock moved to storage
